+++ packages/gatsby-link/src/index.js
@@ -151,12 +151,10 @@ class GatsbyLink extends React.Component {
               }
             }
 
-            // In production, make sure the necessary scripts are
+            // Make sure the necessary scripts and data are
             // loaded before continuing.
-            if (process.env.NODE_ENV === `production`) {
-              e.preventDefault()
-              window.___push(this.state.to)
-            }
+            e.preventDefault()
+            window.___push(this.state.to)
           }
 
           return true