+++ packages/gatsby/src/internal-plugins/query-runner/query-watcher.js
@@ -70,27 +70,9 @@ const handleQuery = (
   query,
   component
 ) => {
-  // If this is page query
-  if (components.has(component)) {
-    if (components.get(component).query !== query.text) {
-      boundActionCreators.replaceComponentQuery({
-        query: query.text,
-        componentPath: component,
-      })
-
-      debug(
-        `Page query in ${component} ${
-          components.get(component).query.length === 0
-            ? `was added`
-            : `has changed`
-        }.`
-      )
-      queueQueriesForPageComponent(component)
-    }
-    return true
-
-    // Add action / reducer + watch staticquery files
-  } else if (query.isStaticQuery) {
+  // If this is a static query
+  // Add action / reducer + watch staticquery files
+  if (query.isStaticQuery) {
     const isNewQuery = !staticQueryComponents.has(query.jsonName)
     if (
       isNewQuery ||
@@ -113,6 +95,25 @@ const handleQuery = (
 
       boundActionCreators.deleteComponentsDependencies([query.jsonName])
       queueQueryForPathname(query.jsonName)
+    }
+    return true
+
+    // If this is page query
+  } else if (components.has(component)) {
+    if (components.get(component).query !== query.text) {
+      boundActionCreators.replaceComponentQuery({
+        query: query.text,
+        componentPath: component,
+      })
+
+      debug(
+        `Page query in ${component} ${
+          components.get(component).query.length === 0
+            ? `was added`
+            : `has changed`
+        }.`
+      )
+      queueQueriesForPageComponent(component)
     }
     return true
   }
@@ -141,10 +142,12 @@ const updateStateAndRunQueries = isFirstRun => {
 
     if (queriesWillNotRun) {
       report.log(report.stripIndent`
-        Queries are only executed for Page components. Instead of a query,
-        co-locate a GraphQL fragment and compose that fragment into the query (or other
-        fragment) of the top-level page that renders this component. For more
-        info on fragments and composition see: http://graphql.org/learn/queries/#fragments
+        Exported queries are only executed for Page components. Instead of an exported
+        query, either co-locate a GraphQL fragment and compose that fragment into the
+        query (or other fragment) of the top-level page that renders this component, or
+        use a <StaticQuery> in this component. For more info on fragments and
+        composition, see http://graphql.org/learn/queries/#fragments and for more
+        information on <StaticQuery>, see https://next.gatsbyjs.org/docs/static-query
       `)
     }
     runQueuedQueries()