+++ packages/gatsby/src/redux/actions.js
@@ -1,6 +1,6 @@
 // @flow
-import Joi from "joi"
-import chalk from "chalk"
+const Joi = require(`joi`)
+const chalk = require(`chalk`)
 const _ = require(`lodash`)
 const { bindActionCreators } = require(`redux`)
 const { stripIndent } = require(`common-tags`)
@@ -12,8 +12,8 @@ const { hasNodeChanged, getNode } = require(`./index`)
 const { trackInlineObjectsInRootNode } = require(`../schema/node-tracking`)
 const { store } = require(`./index`)
 const fileExistsSync = require(`fs-exists-cached`).sync
-import * as joiSchemas from "../joi-schemas/joi"
-import { generateComponentChunkName } from "../utils/js-chunk-names"
+const joiSchemas = require(`../joi-schemas/joi`)
+const { generateComponentChunkName } = require(`../utils/js-chunk-names`)
 
 const actions = {}
 
@@ -1105,6 +1105,30 @@ actions.createRedirect = ({
       toPath: `${pathPrefix}${toPath}`,
       ...rest,
     },
+  }
+}
+
+/**
+ * Add a third-party schema to be merged into main schema. Schema has to be a
+ * graphql-js GraphQLSchema object.
+ *
+ * This schema is going to be merged as-is. This can easily break the main
+ * Gatsby schema, so it's user's responsibility to make sure it doesn't happen
+ * (by eg namespacing the schema).
+ *
+ * @param {Object} $0
+ * @param {GraphQLSchema} $0.schema GraphQL schema to add
+ */
+actions.addThirdPartySchema = (
+  { schema }: { schema: GraphQLSchema },
+  plugin: Plugin,
+  traceId?: string
+) => {
+  return {
+    type: `ADD_THIRD_PARTY_SCHEMA`,
+    plugin,
+    traceId,
+    payload: schema,
   }
 }