+++ packages/gatsby/src/schema/build-node-types.js
@@ -34,7 +34,10 @@ module.exports = async ({ parentSpan }) => {
   const spanArgs = parentSpan ? { childOf: parentSpan } : {}
   const span = tracer.startSpan(`build schema`, spanArgs)
 
-  const types = _.groupBy(getNodes(), node => node.internal.type)
+  const types = _.groupBy(
+    getNodes().filter(node => node.internal && !node.internal.ignoreType),
+    node => node.internal.type
+  )
   const processedTypes: TypeMap = {}
 
   clearTypeExampleValues()
@@ -147,7 +150,6 @@ module.exports = async ({ parentSpan }) => {
 
     const fieldsFromPlugins = await apiRunner(`setFieldsOnGraphQLNodeType`, {
       type: intermediateType,
-      allNodes: getNodes(),
       traceId: `initial-setFieldsOnGraphQLNodeType`,
       parentSpan: span,
     })