+++ packages/gatsby-plugin-catch-links/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.2-beta.5"></a>
+
+## [2.0.2-beta.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-catch-links/compare/gatsby-plugin-catch-links@2.0.2-beta.4...gatsby-plugin-catch-links@2.0.2-beta.5) (2018-08-01)
+
+**Note:** Version bump only for package gatsby-plugin-catch-links
+
+<a name="2.0.2-beta.4"></a>
+
+## [2.0.2-beta.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-catch-links/compare/gatsby-plugin-catch-links@2.0.2-beta.3...gatsby-plugin-catch-links@2.0.2-beta.4) (2018-07-21)
+
+**Note:** Version bump only for package gatsby-plugin-catch-links
+
+<a name="2.0.2-beta.3"></a>
+
+## [2.0.2-beta.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-catch-links/compare/gatsby-plugin-catch-links@2.0.2-beta.2...gatsby-plugin-catch-links@2.0.2-beta.3) (2018-07-17)
+
+### Bug Fixes
+
+- catch-links should use gatsby for push ([#6477](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-catch-links/issues/6477)) ([a3b5b2d](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-catch-links/commit/a3b5b2d))
+
 <a name="2.0.2-beta.2"></a>
 
 ## [2.0.2-beta.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-catch-links/compare/gatsby-plugin-catch-links@2.0.2-beta.1...gatsby-plugin-catch-links@2.0.2-beta.2) (2018-06-20)